% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{histograms}
\alias{histograms}
\title{Histograms}
\usage{
histograms(x)
}
\arguments{
\item{x}{Data set}
}
\value{
A histogram for every numerical column of your data.
}
\description{
Histograms of your data
}
\details{
Histograms are a basic way to understand numerical data. This function provides the base histogram for every numerical column.
}
\author{
Kyle
}
