% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro.R
\name{intro}
\alias{intro}
\title{Introduction}
\usage{
intro(x)
}
\arguments{
\item{x}{Data set}
}
\value{
A readable sentence that introduces your data set.
}
\description{
Introduction to your data
}
\details{
Introduction provides a readable basic overview of your data in a way that is understandable without ever seeing a data set before.
}
\author{
Kyle
}
