% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro_columns.R
\name{intro_columns}
\alias{intro_columns}
\title{intro_columns}
\usage{
intro_columns(x)
}
\arguments{
\item{x}{Data set}
}
\value{
A readable sentence that introduces you to each column of your data set.
}
\description{
Introduces you to the columns of your data set
}
\details{
Intro_columns provides a readable basic overview of the columns of your data set in a way that is understandable and conversational. This is a good second function to run after the intro function when you haven not seen this data set before. It can also help with writing a report.
}
\author{
Kyle Duffy
}
