% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{rows_with_missing_values}
\alias{rows_with_missing_values}
\title{Easily calculates number of rows that have any column with a missing value}
\usage{
rows_with_missing_values(x)
}
\arguments{
\item{x}{Data set}
}
\value{
The amount of rows in your data set that have any missing values in any columns.
}
\description{
Missing values can cause a lot of problems later in analysis. So it is important to know when you get a data set how many are missing. With this function, it is very easy to know how many rows have any column with a missing value.
}
\author{
Kyle
}
